requirements.txt
anyio==3.7.0
certifi==2023.5.7
click==8.1.3
coverage==7.2.7
dnspython==2.3.0
exceptiongroup==1.1.1
fastapi==0.95.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
mongo==0.2.0
motor==3.1.2
packaging==23.1
pluggy==1.0.0
pydantic==1.10.8
pymongo==4.3.3
pytest==7.3.1
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.6.2
uvicorn==0.22.0
